Rename annotations package to confighash
