+++ operator/internal/common/confighash/keys.go
@@ -1,8 +1,8 @@
-package annotations
+package confighash
 
 const (
-	// ConfigHashAnnotation is a level-based annotation that is persisted in the Central and
+	// AnnotationKey is a level-based annotation that is persisted in the Central and
 	// SecuredCluster Deployments and DaemonSets pod templates, to trigger pod restarts when config changes.
 	// Currently only the CA certificate is hashed.
-	ConfigHashAnnotation = "app.stackrox.io/config-hash"
+	AnnotationKey = "app.stackrox.io/config-hash"
 )
